requirements.txt
networkx>=3.2
scikit-learn>=1.4
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.6
pyyaml>=6.0
numpy>=1.26
scipy>=1.12
python-multipart>=0.0.9
httpx>=0.27          # for FastAPI test client
pytest>=8.0
pytest-asyncio>=0.23
torch>=2.0
torch_geometric>=2.4
